Decimal Refactor part 3 Money Migration Baseline

+++ lib/screens/accounts/account_deatails/data/hive_account_details_repository.dart
@@ -51,10 +51,10 @@ class HiveAccountDetailsRepository implements AccountDetailsRepository {
     );
 
     return AccountProjection(
-      balance: projection.balance,
-      available: projection.available,
-      reserved: projection.reserved,
-    );
+  balance: projection.balance.toDouble(),
+  available: projection.available.toDouble(),
+  reserved: projection.reserved.toDouble(),
+);
   }
 
   @override
